refactor: lift task-row minting out of core.ts
core.ts was already 2 over the cap on main and the land preflight pushed it to
516. createTask/openDirectoryTask/adoptScratchRepo were the one cluster in it
that was real logic rather than a delegation to a coordinator, so that is where
it splits: they move to core-create.ts as free functions over the store, and
the class keeps three thin delegators. 516 -> 426.
